% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkTrend_rcpp.R
\name{acf.fft}
\alias{acf.fft}
\title{faster autocorrelation based on ffw}
\usage{
acf.fft(x, lag.max = NULL)
}
\description{
This function is 4-times faster than \code{\link[stats:acf]{stats::acf()}}
}
\references{
\enumerate{
\item https://gist.github.com/FHedin/05d4d6d74e67922dfad88038b04f621c
\item https://gist.github.com/ajkluber/f293eefba2f946f47bfa
\item http://www.tibonihoo.net/literate_musing/autocorrelations.html#wikispecd
\item https://lingpipe-blog.com/2012/06/08/autocorrelation-fft-kiss-eigen
}
}
\keyword{internal}
