% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_plot.R
\name{add_gridLine}
\alias{add_gridLine}
\title{add_gridLine}
\usage{
add_gridLine(dates, col = "grey60", lty = 3, ...)
}
\arguments{
\item{dates}{Date vector}

\item{...}{others to \code{\link[graphics:abline]{graphics::abline()}}}
}
\description{
add_gridLine
}
