% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{addin_cutLines}
\alias{addin_cutLines}
\alias{addin_copyLines}
\title{Cut lines as sublime}
\usage{
addin_cutLines()

addin_copyLines(output = FALSE)
}
\arguments{
\item{output}{Boolean, whether return selection info?}
}
\description{
Cut lines as sublime
}
\keyword{internal}
