% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sp2.R
\name{area.spatial}
\alias{area.spatial}
\title{Calculate area of spatial object}
\usage{
area.spatial(grid, area.weighted = TRUE)
}
\arguments{
\item{grid}{SpatialPolygonsDataFrame or SpatialGridDataFrame}

\item{area.weighted}{if not, ones vector will be return}
}
\description{
Calculate area of spatial object
}
