% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_data.table.R
\name{array2dt}
\alias{array2dt}
\alias{dt2array}
\title{Convert array to data.table}
\usage{
array2dt(arr, dimnames)

dt2array(dt, value_col = "value")
}
\arguments{
\item{arr}{input array}

\item{dimnames}{list of dimension names, e.g., \code{dimnames(arr)}}
}
\value{
data.table
}
\description{
Convert array to data.table
}
\examples{
library(Ipaper)

arr <- array(1:6,
  dim = c(2, 3),
  dimnames = list(
    site = c("A", "B"),
    date = c("d1", "d2", "d3")
    # var = c("v1", "v2", "v3", "v4")
  )
)

# array -> dt
dt <- array2dt(arr, dimnames(arr))
print(dt)

# dt -> array
arr2 <- dt2array(dt)
all.equal(arr, arr2)
}
