% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_sp.R
\name{cdo_grid}
\alias{cdo_grid}
\title{get regular cdo grid}
\usage{
cdo_grid(range = c(70, 140, 15, 55), cellsize = 0.1, outfile = "grid.txt")
}
\arguments{
\item{range}{\verb{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{double}

\item{outfile}{character, path of output file}
}
\description{
get regular cdo grid
}
