% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tag.R
\name{char2script}
\alias{char2script}
\alias{code_ChrVec}
\title{generate R script of character vector}
\usage{
char2script(x, collapse = "\\"", verbose = TRUE)

code_ChrVec(x, collapse = "\\"", verbose = TRUE)
}
\arguments{
\item{x}{character vector, data.frame or list.}

\item{collapse}{an optional character string to separate the results. Not NA_character_.}
}
\description{
generate R script of character vector
}
