% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp.R
\name{clamp}
\alias{clamp}
\alias{clamp_min}
\alias{clamp_max}
\title{clamp}
\usage{
clamp(x, lims = c(0, 1), fill.na = FALSE)

clamp_min(x, value = 0)

clamp_max(x, value = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{lims}{limits}

\item{fill.na}{If true, values of lims are set to NA; else, values are just
constrained in the range of \code{lims}.}
}
\description{
clamp values in the range of \code{lims}
}
\examples{
clamp(1:10, lims = c(4, 7), fill.na = TRUE) 
}
