% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{code_ChrVec}
\alias{code_ChrVec}
\title{generate R script of character vector}
\usage{
code_ChrVec(x, collapse = "\\"")
}
\arguments{
\item{x}{character vector, data.frame or list.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}
}
\description{
generate R script of character vector
}
