% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_func.R
\name{code_editor}
\alias{code_editor}
\alias{subl}
\alias{code}
\alias{smerge}
\alias{SumatraPDF}
\title{code editor}
\usage{
subl(path = getwd())

code(path = getwd())

smerge(path = getwd())

SumatraPDF(path = getwd())
}
\description{
sublime text3 ad vscode
}
\keyword{internal}
