% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_colorkey.R
\name{colorkey_pos}
\alias{colorkey_pos}
\title{colorkey_pos}
\usage{
colorkey_pos(space, ncol = 3)
}
\arguments{
\item{space}{One of \code{c('top', 'bottom', 'right', 'left')}}

\item{ncol}{How many columns in right mode? This parameter is designed for
future extension.}
}
\description{
colorkey_pos
}
\keyword{internal}
