% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.polygon.R
\name{coord.polygon}
\alias{coord.polygon}
\alias{coord.polygons}
\title{Get coordinates of spatialPolygons object}
\usage{
coord.polygon(x)

coord.polygons(x)
}
\arguments{
\item{x}{spatialPolygons object}
}
\value{
\code{lst_loc}
}
\description{
Get coordinates of spatialPolygons object

coord.polygons
}
