% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_levels.R
\name{cut_levels}
\alias{cut_levels}
\title{cut_levels}
\usage{
cut_levels(x, pvalue = c(0.01, 0.05, 0.1), verbose = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{pvalue}{p <= \code{x\%}, means its significant at \code{x\%} level}
}
\description{
cut_levels
}
\examples{
x <- c(-0.09, -0.4, 0.04, 0.15)
cut_levels(x, verbose = TRUE)
}
