% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cut_plevels}
\alias{cut_plevels}
\title{cut_plevels}
\usage{
cut_plevels(x, pvalue = c(0.01, 0.05, 0.1), verbose = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{pvalue}{p <= \verb{x\%}, means its significant at \verb{x\%} level}
}
\description{
cut_plevels
}
\examples{
x <- c(-0.09, -0.4, 0.04, 0.15)
cut_plevels(x, verbose = TRUE)
}
