% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{df2sp}
\alias{df2sp}
\title{df2sp
Convert dataframe data into SpatialPointsDataframe}
\usage{
df2sp(d, formula = ~lon + lat, prj)
}
\arguments{
\item{d}{A data.frame with coordinates information}

\item{formula}{~longitude+latitude}

\item{prj}{\code{\link[sp]{CRS-class}} Projection.}
}
\description{
df2sp
Convert dataframe data into SpatialPointsDataframe
}
