% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dir.show}
\alias{dir.show}
\title{dir.show}
\usage{
dir.show(path = getwd())
}
\description{
open assign path in windows explorer, and default path is 
current directory. This function is only designed for windows system.
}
