% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ipaper.R
\name{download_aria2}
\alias{download_aria2}
\title{download using aria2}
\usage{
download_aria2(DOIs, journal = ".", srcFUN = NULL, n = 8,
  Rshell = FALSE, ...)
}
\arguments{
\item{DOIs}{according to doi, it find corresponding paper and download it. 
such as "10.1175/JHM-D-15-0157.1". URLencoding format, like 
"10.1175\%2FJHM-D-15-0157.1" is also support.}

\item{journal}{journal name, string used to make new directory to save papers}

\item{srcFUN}{function used to generate pdf download links according to DOIs. 
If srcFUN is null, then this function will treat input parameter DOIs as download urls.}

\item{n}{maximum number of parallel downloads. We used aria2c to download 
papers batchly. In this function we set --max-concurrent-downloads, --split and  
--max-connection-per-server all equal to n. Detailed information about this
parameters could be find in aria2c document \url{https://aria2.github.io/manual/en/html/}.}

\item{Rshell}{whether execute aria2c command in R shell. If false, it will save
command string to clipboard, and then you can paste this command in cmd OR other 
shells which can execute aria2c command}

\item{...}{other parameter pass to srcFUN.}
}
\description{
download papers batchly using aria2. `srcFUN` function is used 
to generate pdf download links. You can also construct src function personally 
to support more.
}
\examples{
# First, you need to get doi;
# Second, you need to select the suitable srcFUN of corresponding database. If
#   this package have not yet, you can consider to extend the srcFUN, or 
#   contact me directly.

DOIs <- rep("10.1175\%2FJHM-D-15-0157.1", 4) #test aria2 parallel download
download_aria2(DOIs, journal = "JHM", srcFUN = src_AMS, n = 4, Rshell = TRUE)
}
\author{
Dongdong Kong \url{kongdd@live.cn}
}
