% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_download.R
\name{download_httr}
\alias{download_httr}
\title{download using httr}
\usage{
download_httr(DOIs, journal = ".", srcFUN = NULL, ...)
}
\arguments{
\item{DOIs}{Character vectors, multiple \code{doi}.}

\item{journal}{journal name, string used to make new directory to save papers}

\item{srcFUN}{function used to generate pdf download links according to DOIs. 
If srcFUN is null, then this function will treat input parameter DOIs as download urls.}

\item{...}{other parameter pass to srcFUN.}
}
\description{
Download using httr package by hadley
}
\examples{
# First, you need to get doi;
# Second, you need to select the suitable srcFUN of corresponding database. If
#   this package have not yet, you can consider to extend the srcFUN, or 
#   contact me directly.

\dontrun{
download_httr("10.1175\%2FJHM-D-15-0157.1", journal = '.', srcFUN = src_AMS)
}
}
