% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.colorkey.R
\name{draw.colorkey}
\alias{draw.colorkey}
\title{draw.colorkey}
\usage{
draw.colorkey(key, draw = FALSE, vp = NULL)
}
\arguments{
\item{key}{ A list determining the key. See documentation for
    \code{levelplot}, in particular the section describing the
    \code{colorkey} argument, for details.
  }

\item{draw}{ logical, whether the grob is to be drawn. }

\item{vp}{ viewport }
}
\description{
draw.colorkey
}
\examples{
library(grid)
brks_SOS  <- c(-Inf, seq(110, 150, 5), Inf)
key <- list(
    at = brks_SOS, 
    space = 'right', height = 1,
    tri.upper = 0.05,  tri.lower = 0.05, 
    labels=list(cex=1.2, fontface='bold'), 
    rect = list(col = "black", lwd = 0.2)
)
g <- draw.colorkey(key)
grid.newpage()
grid.draw(g)

## sp
library(sp)
demo(meuse, ask = FALSE, echo = FALSE)
spplot(meuse, c("ffreq"), col.regions= "black", 
       pch=c(1,2,3), key.space=list(x=0.1,y=.95,corner=c(0,1)))
spplot(meuse.grid)
}
