% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{dt_day2year}
\alias{dt_day2year}
\title{dt_day2year}
\usage{
dt_day2year(
  dat,
  nmiss_day_per_mon = 3,
  nmiss_MonPerYear = 0,
  nmin_year = 55,
  ...
)
}
\arguments{
\item{dat}{A data.table, at least with the columns of \code{c("site", "date", "value")}}
}
\description{
dt_day2year
}
