% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_dplyr.R
\name{dt_tools}
\alias{dt_tools}
\alias{dt_round}
\alias{dt_chr2num}
\title{data.frame manipulating function by \code{dplyr::across}}
\usage{
dt_round(d, digits = 2)

dt_chr2num(d, fun = as.numeric)
}
\description{
data.frame manipulating function by \code{dplyr::across}
}
