% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_envelope.R
\name{ensemble_mean}
\alias{ensemble_mean}
\title{ensemble_mean}
\usage{
ensemble_mean(x, y, step = 0.2, chunk = NULL, FUN = "mean")
}
\arguments{
\item{FUN}{function of mean or median}
}
\description{
ensemble_mean
}
