% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extractId}
\alias{extractId}
\title{extractId}
\usage{
extractId(sp, shpfile)
}
\arguments{
\item{sp}{A SpatialPointDataFrame with the station coordinates information}

\item{shpfile}{A character, shape file path.}
}
\description{
extractId
}
