% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name.R
\name{file_ext}
\alias{file_ext}
\alias{file_name}
\title{get file name and extension}
\usage{
file_ext(file)

file_name(file)
}
\arguments{
\item{file}{file path}
}
\description{
get file name and extension
}
\examples{
file_name("./a.pdf")
file_ext("./a.pdf")
}
