% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fprintf}
\alias{fprintf}
\title{fprintf
Print sprintf result into console, just like C style fprintf function}
\usage{
fprintf(fmt, ...)
}
\arguments{
\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{other parameters will be passed to \code{sprintf}}
}
\description{
fprintf
Print sprintf result into console, just like C style fprintf function
}
\examples{
cat(fprintf("\%s\n", "Hello phenofit!"))
}
