% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_data.table.R
\name{fread_dir}
\alias{fread_dir}
\title{fread_dir}
\usage{
fread_dir(indir, pattern = "*.csv", ..., .progress = "text", list2df = FALSE)
}
\arguments{
\item{...}{others to \code{\link[data.table:fread]{data.table::fread()}}}
}
\description{
fread_dir
}
