% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colorkey.R
\name{get_colorkey}
\alias{get_colorkey}
\title{Construct lattice colorkey}
\usage{
get_colorkey(
  breaks,
  cols = NULL,
  space = "bottom",
  lgd.title = NULL,
  is_factor = FALSE,
  unit = NULL,
  unit.adj = 0.3,
  cex = 1.3,
  fontfamily = "Times",
  fontface = 2
)
}
\description{
Construct lattice colorkey
}
