% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ipaper.R
\name{get_DOIs}
\alias{get_DOIs}
\title{get DOIs from endnote export xml files}
\usage{
get_DOIs(file)
}
\arguments{
\item{file}{Endnote exported xml file path, DOI information must be included.}
}
\description{
get DOIs from endnote export xml files
}
