% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_grid}
\alias{get_grid}
\title{get_grid
construct SpatialGridDataFrame}
\usage{
get_grid(range, cellsize, midgrid = c(TRUE, TRUE), prj = prj84)
}
\arguments{
\item{range}{A numeric vector, [lat_min, lat_max, lon_min, lon_max]}

\item{cellsize}{A numeric, grid cell size.}

\item{midgrid}{A vector, [midgrid_long, midgrid_lat]. If midgrid = false, then
begin point and end point locate on grid lines; If true, then begin point and
end point in the middle of grid.}

\item{prj}{\code{\link[sp]{CRS-class}} Projection.}
}
\description{
get_grid
construct SpatialGridDataFrame
}
\examples{
range <- c(25, 40, 73, 105) # Tibetan Plateau
Grid  <- get_grid(range, cellsize = 1/12, midgrid = TRUE)
}
