% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_nearGrids}
\alias{get_nearGrids}
\title{get_nearGrids}
\usage{
get_nearGrids(grid, station, ngrid = 3 * 3)
}
\arguments{
\item{grid}{SpatialGridDataFrame object}

\item{ngrid}{how many nearest grids to be extracted?}

\item{sp}{SpatialPointDataframe object}
}
\value{
nearId nearest order
}
\description{
find nearest 3*3 grids of points
}
\details{
with the help of RANN package to quickly select nearest 3*3 grids
use line dist to replace sphere distance
}
