% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart2.R
\name{get_perc.factor}
\alias{get_perc.factor}
\title{get factor frequency}
\usage{
get_perc.factor(z, w = NULL)
}
\description{
get factor frequency
}
\examples{
\dontrun{
get_perc.factor(factor(c(1, 2, 4), 1:4))
}
}
\keyword{internal}
