% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_download.R
\name{getDOIs_endnote}
\alias{getDOIs_endnote}
\title{get DOIs from endnote export xml files}
\usage{
getDOIs_endnote(xmlfile)
}
\arguments{
\item{xmlfile}{Endnote exported xml file path, DOI information must be included.}
}
\description{
get DOIs from endnote export xml files
}
