% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPointsMODISTile.R
\name{getPointsMODISTile}
\alias{getPointsMODISTile}
\title{Get points data from MODIS tiles}
\usage{
getPointsMODISTile(coor, cellsize = 1/240)
}
\arguments{
\item{coor}{2-column matrix of coordinates (longitude, latitude)}

\item{cellsize}{500m products, cellsize is regarded as 1/240; 1km product, 
cellsize is regarded as 1/120.}
}
\value{
\itemize{
\item{h} MODIS tile horizontal id
\item{v} MODIS tile vertical id
\item{i} point horizontal position at the tile
\item{j} point vertical position at the tile
}
}
\description{
Get points data from MODIS tiles
}
\references{
https://landweb.modaps.eosdis.nasa.gov/cgi-bin/developer/tilemap.cgi
https://modis-land.gsfc.nasa.gov/MODLAND_grid.html
}
