% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getwd_clip}
\alias{getwd_clip}
\title{getwd_clip}
\usage{
getwd_clip()
}
\description{
get directory path in clipboard, same as getwd function
}
\references{
https://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
