% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_tools}
\alias{git_tools}
\alias{git_push}
\alias{git_commit_amend}
\alias{git_commit}
\alias{git_set_remote}
\title{git tools}
\usage{
git_push(f = FALSE)

git_commit_amend()

git_commit(title = Sys.time())

git_set_remote(url)
}
\description{
git tools
}
