% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_apply.R
\name{group_apply}
\alias{group_apply}
\alias{upper_envelope}
\title{group_apply}
\usage{
group_apply(x, y, FUN = "mean", nchunk = 20, step = 0.2, brks = NULL, ...)

upper_envelope(x, y, interval = c(0.5, 0.8, 0.9, 0.95), nchunk = 50)
}
\arguments{
\item{FUN}{function of mean or median}

\item{nchunk}{split data into n chunks by setting \code{brks = pretty(x, nchunk)},
and apply \code{FUN} in every group. If \code{brks} provided, \code{step} will be ignored.}

\item{step}{split data into chunks byn setting \code{brks = seq(xrange[1], xrange[2], by = step)}.
If \code{brks} or \code{nchunk} provided, \code{step} will be ignored.}

\item{brks}{a numeric vector}

\item{...}{others to \code{FUN}}
}
\description{
group_apply
}
\keyword{internal}
