% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_dplyr.R
\name{grouped_list}
\alias{grouped_list}
\title{grouped_list}
\usage{
grouped_list(data, group)
}
\arguments{
\item{data}{A list object, \code{length(data) = nrow(group)}, with elements in the
same order as group rows}

\item{group}{A data.frame or data.table object}
}
\description{
grouped_list
}
