% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{import}
\alias{import}
\title{import data to R}
\usage{
import(path, ..., nthreads = 6)
}
\arguments{
\item{nthreads}{The number of threads to use when reading data (the initial value is 1L). When TBB is not available, values greater than 1 emit a warning and fall back to 1.}
}
\description{
import data to R
}
\details{
Support rda, rds, fst, csv, qs2
}
