% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ipaper.R
\name{Init_Check}
\alias{Init_Check}
\title{check outdir and doi}
\usage{
Init_Check(doi, outdir)
}
\arguments{
\item{doi}{charater or caracter vector are also support. This function also
can also decode URLdecode format doi. So doi, like this 
"10.1175\%2FJHM-D-15-0157.1" is also support.}
}
\value{
URLdecode doi
}
\description{
Check whether outdir exist, if not then will create it. If doi was URLencoded, 
then decode it.
}
