% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_colorkey.R
\name{key_gf}
\alias{key_gf}
\title{key_gf_nonraster}
\usage{
key_gf(key, key.layout, vp, vp_label, reccentre, recdim, border = TRUE, ...)
}
\arguments{
\item{...}{ignored.}
}
\description{
key_gf_nonraster
}
\examples{
\dontrun{
key.gf <- key_gf(key, key.layout, vp, vp_label, axis.line, reccentre, recdim)
}
}
\keyword{internal}
