% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_colorkey.R
\name{key_label}
\alias{key_label}
\title{key_label}
\usage{
key_label(key.gf, key, labscat, labelsGrob, vp_label, axis.line, ...)
}
\description{
key_label
}
\examples{
\dontrun{
key.gf <- key_label(key.gf, key, labscat, vp_label, axis.line)
}
}
\keyword{internal}
