% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tag.R
\name{label_tag}
\alias{label_tag}
\alias{char2expr}
\title{label_tag}
\usage{
label_tag(labels, tag = TRUE, expression = TRUE, letter_begin = 1)

char2expr(labels)
}
\arguments{
\item{labels}{character vector or expression vector}

\item{tag}{boolean}
}
\description{
label_tag
}
\examples{
label_tag(1:5)
char2expr(1:5)
}
