% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listk.R, R/melt_list.R
\name{listk}
\alias{listk}
\alias{melt_list}
\title{listk}
\usage{
listk(...)

melt_list(list, var.name, na.rm = TRUE, ...)
}
\arguments{
\item{...}{objects, possibly named.}

\item{list}{A list object, with the same colnames data.frame in every element.}

\item{var.name}{vector of id variables. Can be integer (variable position)
or string (variable name). If blank, will use all non-measured variables.}

\item{na.rm}{Boolean}

\item{...}{other parameters to melt}
}
\description{
listk

melt_list
}
\examples{
a = 1
b = 1:2
c = 1:3
l1 <- listk(a, b, c)
l2 <- listk(a, b, c = 1:3)
l3 <- listk(a = 1, b = 1:2, c = 1:3)
l4 <- listk(1, 1:2, c)
# data.frame
df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
l  <- list(a = df, b = df)
df_new <- melt_list(l, "id")

# data.table
df <- data.table::data.table(year = 2010, day = 1:3, month = 1, site = "A")
l  <- list(a = df, b = df)
df_new <- melt_list(l, "id")
}
