% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVIDEO.R
\name{makeVIDEO}
\alias{makeVIDEO}
\title{makeVIDEO}
\usage{
makeVIDEO(file = "ffmpeg_\%d.avi", path = ".", pattern = "*.png",
  mode = c("ultrafast", "slow", "veryslow"))
}
\arguments{
\item{file}{filename of output video}

\item{path}{The path of figures, i.e, pngs, to make video}

\item{pattern}{regular expression pattern to filter figures}

\item{mode}{must be one of "ultrafast", "slow", "veryslow"}
}
\description{
make video through ffmpeg
}
