% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_list.R
\name{melt_list}
\alias{melt_list}
\alias{melt_tree}
\title{melt_list}
\usage{
melt_list(list, var.name = "variable", na.rm = TRUE, ...)

melt_tree(x, names, ...)
}
\arguments{
\item{list}{A list object, with the same colnames data.frame in every element.}

\item{var.name}{vector of id variables. Can be integer (variable position)
or string (variable name). If blank, will use all non-measured variables.}

\item{na.rm}{Boolean}

\item{...}{other parameters to melt}
}
\description{
melt_list
}
\examples{
# data.frame
df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
l <- list(a = df, b = df)
df_new <- melt_list(l, "id")

# data.table
df <- data.table::data.table(year = 2010, day = 1:3, month = 1, site = "A")
l <- list(a = df, b = df)
df_new <- melt_list(l, "id")
}
