% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_terminal.R
\name{ok}
\alias{ok}
\alias{warn}
\alias{num_bad}
\alias{num_good}
\title{Colored terminal output}
\usage{
ok(...)

warn(...)

num_bad(str, width = NULL, ...)

num_good(str, width = NULL, ...)
}
\arguments{
\item{...}{Strings to style.}
}
\description{
Colored terminal output
}
\keyword{internal}
