% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pal}
\alias{pal}
\title{pal
show colors in figure device}
\usage{
pal(col, border = "light gray")
}
\arguments{
\item{col}{colors to be visualize.}

\item{border}{rect border for each color}
}
\description{
pal
show colors in figure device
}
