% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.annotation.R
\name{panel.annotation}
\alias{panel.annotation}
\title{panel.annotation}
\usage{
panel.annotation(grob, bbox = c(0, 0.5, 0.5, 1), ...)
}
\arguments{
\item{grob}{grid object}

\item{bbox}{\code{[xmin, xmax, ymin, ymax]}}

\item{...}{ignored}
}
\description{
panel.annotation
}
