% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.shade.R
\name{panel.shade}
\alias{panel.shade}
\alias{panel.shade.list}
\alias{panel.poly_grid}
\alias{panel.polygonsplot2}
\alias{panel.gridplot2}
\title{panel.shade}
\usage{
panel.shade(
  x,
  y = NULL,
  density = NULL,
  angle = 45,
  border = NULL,
  col = NA,
  lty = par("lty"),
  fill = NA,
  fillOddEven = FALSE,
  ...
)

panel.shade.list(lst, ...)

panel.poly_grid(
  s,
  union = TRUE,
  density = 0.1,
  angle = 45,
  col = "black",
  lwd = 1,
  lty = 1,
  border = NULL,
  ...
)

panel.polygonsplot2(
  x,
  y,
  z,
  subscripts,
  ...,
  union = FALSE,
  density = 0.2,
  angle = 45,
  col = "grey65",
  lwd = 0.5,
  lty = 1,
  poly_shade,
  sp.layout
)

panel.gridplot2(
  x,
  y,
  z,
  subscripts,
  ...,
  union = FALSE,
  density = 0.2,
  angle = 45,
  col = "grey65",
  lwd = 0.5,
  lty = 1,
  poly_shade,
  sp.layout
)
}
\arguments{
\item{x}{vectors containing the coordinates of the vertices
    of the polygon.}

\item{y}{vectors containing the coordinates of the vertices
    of the polygon.}

\item{density}{the density of shading lines, in lines per inch.  The
    default value of \code{NULL} means that no shading lines are drawn.
    A zero value of \code{density} means no shading nor filling whereas
    negative values and \code{NA} suppress shading (and so allow
    color filling).}

\item{angle}{the slope of shading lines, given as an angle
     in degrees (counter-clockwise).}

\item{border}{the color to draw the border.  The default, \code{NULL},
    means to use \code{\link[graphics]{par}("fg")}.  Use \code{border = NA} to
    omit borders.

    For compatibility with S, \code{border} can also be logical, in
    which case \code{FALSE} is equivalent to \code{NA} (borders omitted)
    and \code{TRUE} is equivalent to \code{NULL} (use the foreground colour),
  }

\item{col}{the color for filling the polygon. The default,
     \code{NA}, is to leave polygons unfilled, unless \code{density} is
     specified.  (For back-compatibility, \code{NULL} is equivalent to
     \code{NA}.)  If \code{density} is specified with a positive value
     this gives the color of the shading lines.}

\item{lty}{the line type to be used, as in \code{\link[graphics]{par}}.}

\item{fillOddEven}{logical controlling the polygon shading mode:  see
      below for details.  Default \code{FALSE}.}

\item{...}{\itemize{
\item \code{panel.shade}     : other parameters to \code{polygon.fullhatch};
\item \code{panel.shade.list}: other parameters to \code{\link[=panel.shade]{panel.shade()}}
}}

\item{lst}{List of coordinates matrix \verb{[x, y]}}
}
\description{
panel.shade
}
\examples{
\dontrun{

library(sp2)
library(sp)
# library(grid)

range <- c(70, 140, 15, 55) # Tibetan Plateau
r  <- get_grid(range, cellsize = 1/2, midgrid = TRUE)

# set.seed(1)
I <- c(1:2000, 4000:6000)
poly   <- as(r[I, ], "SpatialPolygonsDataFrame")
poly2  <- union_polygon(poly)

print(spplot(poly2 , poly_shade = poly2, panel = panel.polygonsplot2, fill = "grey", col = "red"))
print(spplot(r[I, ], poly_shade = poly2, panel = panel.gridplot2    , fill = "grey", col = "red"))
}
}
\seealso{
\link[graphics:polygon]{graphics::polygon}
}
\author{
XiHui Gu
}
