% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot_grid.R
\name{panel.spatial}
\alias{panel.spatial}
\title{panel_hist}
\usage{
panel.spatial(x, y, z, subscripts, ..., contour = FALSE, grob = NULL,
  bbox, sub.hist = TRUE, sp.layout, pars, class = NULL,
  interpolate = TRUE, df.pvalue = NULL, SpatialPixels = NULL,
  data.stat = NULL)
}
\arguments{
\item{par}{position of title, \code{list(title = list(x, y))}}
}
\description{
panel_hist
}
