% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot_grid.R
\name{panel_hist}
\alias{panel_hist}
\title{panel_hist}
\usage{
panel_hist(x, y, z, subscripts, ..., contour = FALSE, grob = NULL,
  bbox, sub.hist = TRUE, sp.layout, pars, class = NULL,
  interpolate = TRUE, data.stat = NULL)
}
\arguments{
\item{par}{position of title, \code{list(title = list(x, y))}}
}
\description{
panel_hist
}
