% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_sbatch.R
\name{par_sbatch}
\alias{par_sbatch}
\title{par_sbatch}
\usage{
par_sbatch(X, FUN, ..., return.res = F, Save = F, outdir = ".")
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector. Other objects
(including classed objects) will be coerced by as.list.}

\item{FUN}{the function to be applied to (mclapply) each element of X or
(mcmapply) in parallel to ....}

\item{...}{other parameters passed to mclapply}

\item{return.res}{Default is no return result.}

\item{Save}{save result in rda file or not?}

\item{outdir}{String, output directory for rda file if Save = true.}
}
\description{
R parallel function, designed for linux sbatch model.
}
