% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_func.R
\name{path.mnt}
\alias{path.mnt}
\title{convert windows path to wsl}
\usage{
path.mnt(path)
}
\arguments{
\item{path}{character vector of file paths.}
}
\description{
In windows system, conversion will not be applied.
}
