% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_func.R
\name{path_mnt}
\alias{path_mnt}
\alias{path_win}
\title{convert windows path to wsl}
\usage{
path_mnt(path)

path_win(path)
}
\arguments{
\item{path}{character vector of file paths.}
}
\description{
In windows system, conversion will not be applied.
}
