% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_pdf.R
\name{pdf_to_emf}
\alias{pdf_to_emf}
\title{Convert PDF to EMF}
\usage{
pdf_to_emf(filename)
}
\arguments{
\item{filename}{A PDF file name or file path}
}
\description{
Convert a PDF file to an EMF file. Requires \code{inkscape} command-line utility.
}
\details{
The function creates an EMF file at the same location where the PDF is.
}
\examples{
\dontrun{
pdf_to_emf(filename = "~/test.pdf")
}
}
\keyword{internal}
