% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_func.R
\name{pdf_view}
\alias{pdf_view}
\alias{SumatraPDF}
\alias{evince}
\title{pdf_view}
\usage{
pdf_view(file, ...)

SumatraPDF(path = getwd(), verbose = FALSE)

evince(path = getwd(), verbose = FALSE)
}
\arguments{
\item{file}{the path of pdf file}
}
\description{
pdf_view
}
\note{
not work in wsl
}
