% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ufile.R
\name{read_excel}
\alias{read_excel}
\title{read csv or xls/xlsx file}
\usage{
read_excel(f, ...)
}
\arguments{
\item{...}{others to fun, one of \code{\link[data.table:fread]{data.table::fread()}}, \code{\link[readxl:read_xls]{readxl::read_xls()}}, \code{\link[=read_xlsx]{read_xlsx()}}}
}
\description{
read csv or xls/xlsx file
}
