% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gee.R
\name{read_jsons.gee}
\alias{read_jsons.gee}
\alias{read_json.gee}
\title{read json files exported from GEE}
\usage{
read_jsons.gee(indir, pattern = "phenoflux212.*.geojson",
  baseprob = c("site", "date"), IsSave = T)

read_json.gee(file, baseprob = c("site", "date"), is_buffer)
}
\arguments{
\item{indir}{Input directory which contains the target json files}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}

\item{baseprob}{the properties name of input point.}

\item{IsSave}{save or not?}

\item{file}{json file path}

\item{is_buffer}{buffered or not?}
}
\description{
read json files exported from GEE

read json file exported from GEE
}
