% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_xlsx2list}
\alias{read_xlsx2list}
\title{read_xlsx2list}
\usage{
read_xlsx2list(file, ...)
}
\arguments{
\item{file}{xlsx or xls file path}

\item{...}{other parameters to \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\description{
If excel file hava many sheets, this function also works.
}
