% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{readxlsx_ToList}
\alias{readxlsx_ToList}
\title{readxlsx_ToList}
\usage{
readxlsx_ToList(file, ...)
}
\arguments{
\item{file}{xlsx or xls file path}

\item{...}{other parameters to \code{\link[readxl]{read_excel}}}
}
\description{
If excel file hava many sheets, this function also works.
}
