% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_terminal.R, R/main_dt_ddply.R,
%   R/reexports.R, R/set_font.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bold}
\alias{red}
\alias{green}
\alias{underline}
\alias{.}
\alias{\%||\%}
\alias{map}
\alias{mapvalues}
\alias{revalue}
\alias{qd_save}
\alias{qd_read}
\alias{select}
\alias{rename}
\alias{arrange}
\alias{mutate}
\alias{top_n}
\alias{top_frac}
\alias{as_tibble}
\alias{tribble}
\alias{group_by}
\alias{group_map}
\alias{\%>\%}
\alias{\%<>\%}
\alias{set_names}
\alias{data.table}
\alias{is.data.table}
\alias{fread}
\alias{fwrite}
\alias{as.data.table}
\alias{make_date}
\alias{make_datetime}
\alias{glue}
\alias{days_in_month}
\alias{\%dopar\%}
\alias{\%do\%}
\alias{iter}
\alias{rowSums2}
\alias{rowMeans2}
\alias{rowQuantiles}
\alias{icount}
\alias{install_github}
\alias{install_gitlab}
\alias{install_git}
\alias{edit_r_environ}
\alias{edit_r_profile}
\alias{use_git_ignore}
\alias{use_build_ignore}
\alias{\%->\%}
\alias{\%<-\%}
\alias{showtext_auto}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon:bold]{bold()}}, \code{\link[crayon:green]{green()}}, \code{\link[crayon:red]{red()}}, \code{\link[crayon:underline]{underline()}}}

  \item{data.table}{\code{\link[data.table:as.data.table]{as.data.table()}}, \code{\link[data.table:data.table]{data.table()}}, \code{\link[data.table:fread]{fread()}}, \code{\link[data.table:fwrite]{fwrite()}}, \code{\link[data.table:is.data.table]{is.data.table()}}}

  \item{dplyr}{\code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:as_tibble]{as_tibble}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:group_map]{group_map()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:top_frac]{top_frac()}}, \code{\link[dplyr:top_n]{top_n()}}, \code{\link[dplyr:tribble]{tribble}}}

  \item{foreach}{\code{\link[foreach:\%do\%]{\%do\%}}, \code{\link[foreach:\%dopar\%]{\%dopar\%}}}

  \item{glue}{\code{\link[glue:glue]{glue()}}}

  \item{iterators}{\code{\link[iterators:icount]{icount()}}, \code{\link[iterators:iter]{iter()}}}

  \item{lubridate}{\code{\link[lubridate:days_in_month]{days_in_month()}}, \code{\link[lubridate:make_date]{make_date()}}, \code{\link[lubridate:make_datetime]{make_datetime()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}, \code{\link[magrittr:set_names]{set_names()}}}

  \item{matrixStats}{\code{\link[matrixStats:rowMeans2]{rowMeans2()}}, \code{\link[matrixStats:rowQuantiles]{rowQuantiles()}}, \code{\link[matrixStats:rowSums2]{rowSums2()}}}

  \item{plyr}{\code{\link[plyr:.]{.()}}, \code{\link[plyr:mapvalues]{mapvalues()}}, \code{\link[plyr:revalue]{revalue()}}}

  \item{purrr}{\code{\link[purrr:\%||\%]{\%||\%}}, \code{\link[purrr:map]{map()}}}

  \item{qs2}{\code{\link[qs2:qd_read]{qd_read()}}, \code{\link[qs2:qd_save]{qd_save()}}}

  \item{remotes}{\code{\link[remotes:install_git]{install_git()}}, \code{\link[remotes:install_github]{install_github()}}, \code{\link[remotes:install_gitlab]{install_gitlab()}}}

  \item{showtext}{\code{\link[showtext:showtext_auto]{showtext_auto()}}}

  \item{usethis}{\code{\link[usethis:edit_r_environ]{edit_r_environ()}}, \code{\link[usethis:edit_r_profile]{edit_r_profile()}}, \code{\link[usethis:use_build_ignore]{use_build_ignore()}}, \code{\link[usethis:use_git_ignore]{use_git_ignore()}}}

  \item{zeallot}{\code{\link[zeallot:\%->\%]{\%->\%}}, \code{\link[zeallot:\%<-\%]{\%<-\%}}}
}}

