% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_jupyter.R
\name{set_jupyter}
\alias{set_jupyter}
\alias{set_dirRoot}
\title{set_jupyter}
\usage{
set_jupyter(width = 10, height = 6, res = 200)

set_dirRoot(verbose = TRUE, dir = getwd())
}
\description{
set_jupyter
}
\references{
\enumerate{
\item https://stackoverflow.com/questions/42729049/how-to-change-the-size-of-r-plots-in-jupyter
}
}
