% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_sp.R
\name{sf_rect}
\alias{sf_rect}
\title{sf_rect}
\usage{
sf_rect(range, crs = st_crs(4326))
}
\description{
sf_rect
}
\examples{
xlim = c(112, 115)
ylim = c(21, 23)
range <- c(xlim, ylim)
poly = sf_rect(range)
write_sf(poly, "poly.shp")

}
