% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fig.R
\name{showfig}
\alias{showfig}
\title{showfig in external app}
\usage{
showfig(file, use.file_show = FALSE)
}
\arguments{
\item{use.file_show}{boolean. If true, \code{file.show} will be used mandatorily.}
}
\description{
showfig in external app
}
\keyword{internal}
