% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\alias{slope_p}
\alias{slope_mk}
\alias{slope_boot}
\title{slope}
\usage{
slope(y, x)

slope_p(y, x)

slope_mk(x)

slope_boot(y, slope_FUN = slope, times = 100, alpha = 0.1, seed)
}
\description{
\itemize{
\item \code{slope}     : linear regression slope
\item \code{slope_p}   : linear regression slope and p-value
\item \code{slope_mk}  : mann kendall Sen's slope and p-value
\item \code{slope_boot}: bootstrap slope
}
}
\examples{
y <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
r    <- slope(y)
r_p  <- slope_p(y)
r_mk <- slope_mk(y)
r_boot <- slope_boot(y)
}
\keyword{internal}
