% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arr_trend.R
\name{slope_arr}
\alias{slope_arr}
\title{slope_arr}
\usage{
slope_arr(arr, FUN = rtrend::slope_mk, return.list = FALSE)
}
\arguments{
\item{arr}{a matrix (\verb{[ngrid, ntime]}) or array (\verb{[nlon, nlat, ntime]}).}

\item{FUN}{slope functions, see \code{\link[rtrend:slope_mk]{slope_mk()}}.}

\item{return.list}{boolean,
\itemize{
\item \code{TRUE}: list(slope, pvalue) will be return
\item \code{FALSE}: a array, with the dim of \verb{[nx, ny, 2]}.
}}
}
\value{
t, A 3d array, with the dim of \verb{[nx, ny, 2]}.
\itemize{
\item \code{t[,,1]}: slope
\item \code{t[,,2]}: pvalue
}
}
\description{
slope_arr
}
