% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot_grid.R
\name{spplot_grid}
\alias{spplot_grid}
\title{Plot methods for spatial data with attributes}
\usage{
spplot_grid(
  grid,
  zcols,
  brks,
  colors,
  col.rev = FALSE,
  toFactor = FALSE,
  sub.hist = TRUE,
  grob = NULL,
  bbox = c(0, 0.5, 0.5, 1),
  panel.title = NULL,
  unit = "",
  unit.adj = 0.3,
  pars = list(title = list(x = 77, y = 39, cex = 1.5), hist = list(origin.x = 77,
    origin.y = 28, A = 15, by = 0.4)),
  stat = list(show = FALSE, name = "RC", loc = c(81.5, 26.5), digit = 1, include.sd =
    FALSE),
  area.weighted = FALSE,
  legend.space = "right",
  layout = NULL,
  colorkey = TRUE,
  interpolate = FALSE,
  lgd.title = NULL,
  sp.layout = NULL,
  par.settings = opt_trellis_default,
  par.settings2 = list(axis.line = list(col = "transparent")),
  ...
)
}
\arguments{
\item{sub.hist}{Boolean. If false, hist will be eliminated.}

\item{pars}{parameters controlling hist, e.g. \code{list(title = list(x=77, y=39, cex=1.5),  hist = list(origin.x=77, origin.y=28, A=15, by = 0.4))}.}

\item{colorkey}{Boolean or list returned by \code{\link[=get_colorkey]{get_colorkey()}}.}

\item{sp.layout}{NULL or list; see notes below}

\item{...}{other parameters to spplot, for example:
\itemize{
\item xlim
\item ylim
\item ...
}}
}
\description{
Lattice (trellis) plot methods for spatial data with attributes
}
\examples{
data("grid_avhrr")

{
    pars = list(title = list(x=77, y=39, cex=1.5), 
                hist = list(origin.x=77, origin.y=28, A=15, by = 0.4))
    stat = list(show = TRUE, name="mean", loc = c(82.5, 26.5), digit = 1, include.sd = TRUE)
    
    p <- spplot_grid(grid_avhrr, 
                     stat = stat, 
                     pars = NULL, 
                     sub.hist = FALSE,
                     interpolate = FALSE)
    p2 <- levelplot2(X1982 ~ s1+s2, grid_avhrr@data, grid_avhrr)
    write_fig(p , "ex-spplot_grid.png", 9.8, 5)
    write_fig(p2, "ex-spplot_grid2.png", 9.8, 5)
}
}
\seealso{
\code{\link[sp:spplot]{sp::spplot()}}, \code{\link[lattice:levelplot]{lattice::levelplot()}}
}
