% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcFUN.R
\name{src_hess}
\alias{src_hess}
\title{srcFUN of hess}
\usage{
src_hess(doi)
}
\arguments{
\item{doi}{DOI like this format "10.5194/hess-20-4191-2016"}
}
\description{
Generate paper of HESS journal pdf download source link
}
