% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcFUN.R
\name{src_SciDirect.doi}
\alias{src_SciDirect.doi}
\title{srcFUN of elsevier database}
\usage{
src_SciDirect.doi(DOIs, outdir = "./", .download = TRUE, ...)
}
\arguments{
\item{DOIs}{according to doi, it find corresponding paper and download it. 
such as "10.1175/JHM-D-15-0157.1". URLencoding format, like 
"10.1175\%2FJHM-D-15-0157.1" is also support.}

\item{outdir}{output file directory}
}
\description{
Just pdf src returned. If you want to download directly you 
can use download_httr(doi, journal = '.', srcFUN = src_SciDirect)
}
