% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcFUN.R
\name{src_SciDirect}
\alias{src_SciDirect}
\title{srcFUN of elsevier database}
\usage{
src_SciDirect(doi)
}
\description{
Just pdf src returned. If you want to download directly you 
can use download_httr(doi, journal = '.', srcFUN = src_SciDirect)
}
