% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcFUN.R
\name{src_SciDirect.url}
\alias{src_SciDirect.url}
\title{srcFUN of elsevier database}
\usage{
src_SciDirect.url(urls, outdir = "./", srcDownload = TRUE, ...)
}
\description{
download pdfs tthrough urls other than DOIs
}
