% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mk.R
\name{stat_mk}
\alias{stat_mk}
\title{stat_mk}
\usage{
stat_mk(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
stat_mk
}
\examples{
library(ggplot2)

ggplot(mpg, aes(displ, hwy, colour = drv)) + 
    geom_point() + 
    stat_mk()
}
