% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_lattice}
\alias{theme_lattice}
\title{theme_lattice}
\usage{
theme_lattice(plot.margin, key.margin, axis.margin, axis.components.inner,
  axis.components.outer, ..., layout.widths, layout.heights, par.settings)
}
\arguments{
\item{plot.margin}{plot padding in layout.heights and layoutheights}

\item{key.margin}{legend margin}

\item{axis.margin}{axis margin}

\item{axis.components.inner, axis.components.outer}{margin object in the order to \code{"top", "right", "bottom", "left"}.}
}
\description{
theme_lattice
}
\note{
key.margin left not work (right controls)
}
