% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_list.R, R/reorder_name.R
\name{rm_empty}
\alias{rm_empty}
\alias{reorder_name}
\alias{contain}
\title{rm_empty}
\usage{
rm_empty(x)

reorder_name(
  d,
  headvars = c("site", "date", "year", "doy", "d8", "d16"),
  tailvars = ""
)

contain(d, pattern = "NDVI|EVI")
}
\arguments{
\item{x}{A vector or list}

\item{d}{A data.frame vector, or list}

\item{headvars}{headvars will be in the head columns.}

\item{tailvars}{tailvars will be in the tail columns.}

\item{pattern}{string used to match \code{names(d)}}
}
\description{
\itemize{
\item \code{reorder_name}: reorder the name of data.frame, date.table or list.
}
}
\examples{
# numeric
x <- c(1:5, NA)
rm_empty(x)

# list
l <- list(1:5, NULL, NA)
rm_empty(l)

df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
dt <- data.table::data.table(year = 2010, day = 1:3, month = 1, site = "A")
l <- list(year = 2010, day = 1:3, month = 1, site = "A")

newname <- c("site", "year")
reorder_name(df, newname)
reorder_name(dt, newname)
reorder_name(l, newname)

df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
contain(df, "year|month|day")
}
\keyword{internal}
