% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_envelope.R
\name{upper_envelope}
\alias{upper_envelope}
\title{function to separate data to steps of x, obtain 95 quantile value for smooth}
\usage{
upper_envelope(x, y, step = 0.2, alpha = 0.95)
}
\description{
function to separate data to steps of x, obtain 95 quantile value for smooth
}
