% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which.na}
\alias{which.na}
\alias{which.notna}
\alias{which.isnull}
\alias{which.notnull}
\title{which functions}
\usage{
which.na(x)

which.notna(x)

which.isnull(x)

which.notnull(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
which functions
}
