% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fig2ps.R
\name{write_fig2ps}
\alias{write_fig2ps}
\title{write_fig2ps}
\usage{
write_fig2ps(ps, lgd, ylab.right, file, width = 10, height, nrow = 6)
}
\arguments{
\item{ps}{A list of ggplot figure objects. And}

\item{lgd}{A grid grob object, legend to show in the bottom.}

\item{ylab.right}{y label title}

\item{width}{inch}

\item{height}{inch}

\item{nrow}{}
}
\description{
Write figure to pages
Subplots xlab and ylab are unified, and only keet singe one.
Currently, only support ggplot figures; And only support arrange figures
into rows (nrow*1).
}
