% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_sp.R
\name{write_sp2rgb}
\alias{write_sp2rgb}
\title{write_sp2rgb}
\usage{
write_sp2rgb(
  grid,
  brks,
  cols,
  file = "sp_rgb.tif",
  mask = NULL,
  col_mask = "transparent"
)
}
\arguments{
\item{grid}{SpatialGridPixels or SpatialGridDataFrame object}

\item{brks}{}

\item{cols}{}

\item{mask}{Boolean vector indicate where to mask}

\item{col_mask}{}
}
\description{
write_sp2rgb
}
\examples{
write_sp2rgb(grid, brks, cols, file = "dem_pearl_rgb.tif")
}
