% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcFUN.R
\name{write_webfile}
\alias{write_webfile}
\title{download web file through web link src}
\usage{
write_webfile(src, outdir = "./", file = NULL, ...)
}
\description{
#http://www.sciencedirect.com/science/article/pii/S0034425701002310/pdfft?md5=2e6dfdaa5b680d49fbe09360b5bed6b4&pid=1-s2.0-S0034425701002310-main.pdf
# has an error: for above url
}
