% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arr_3d.R
\name{array_3dTo2d}
\alias{array_3dTo2d}
\title{array_3dTo2d}
\usage{
array_3dTo2d(array, I_grid)
}
\arguments{
\item{array}{array with the dimension of \verb{[nlon, nlat, ntime]}}

\item{I_grid}{subindex of \verb{[nrow, ncol]}}
}
\value{
\verb{[nlat*nlon, ntime]}
}
\description{
array_3dTo2d
}
