% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_func.R
\name{code_editor}
\alias{code_editor}
\alias{subl}
\alias{code}
\alias{smerge}
\alias{SumatraPDF}
\alias{is_wsl}
\alias{OS_type}
\title{code editor}
\usage{
subl(path = getwd(), verbose = FALSE)

code(path = getwd(), verbose = FALSE)

smerge(path = getwd(), verbose = FALSE)

SumatraPDF(path = getwd(), verbose = FALSE)

is_wsl()

OS_type()
}
\arguments{
\item{verbose}{Boolean. Whether to print command into console?}
}
\description{
sublime text3 ad vscode
}
\keyword{internal}
