% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_func.R
\name{code_editor}
\alias{code_editor}
\alias{is_wsl}
\alias{is_win}
\alias{OS_type}
\alias{subl}
\alias{code}
\alias{smerge}
\title{code editor}
\usage{
is_wsl()

is_win()

OS_type()

subl(path = getwd(), verbose = FALSE)

code(path = getwd(), verbose = FALSE)

smerge(path = getwd(), verbose = FALSE)
}
\arguments{
\item{verbose}{Boolean. Whether to print command into console?}
}
\description{
sublime text3 ad vscode
}
\keyword{internal}
