% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fig.R
\name{dev_open}
\alias{dev_open}
\alias{dev_off}
\title{open device for plot}
\usage{
dev_open(file, width, height, res, use.cairo_pdf = FALSE)

dev_off()
}
\arguments{
\item{file}{file path of output figure}

\item{width}{the width of the device in inches.}

\item{height}{the height of the device in inches.}

\item{res}{The nominal resolution in ppi which will be recorded in the
    bitmap file, if a positive integer.  Also used for units other than
    the default.  If not specified, taken as 72 ppi to set the size of
    text and line widths.}

\item{use.cairo_pdf}{This parameter is for pdf type. whether to use \code{grDevices::cairo_pdf}?
\code{cairo_pdf} supports self defined font, but can not create multiple page pdf.}
}
\description{
open device for plot
}
\keyword{internal}
