% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_func.R
\name{dir.show}
\alias{dir.show}
\title{Open directory in Explorer}
\usage{
dir.show(path = getwd(), verbose = FALSE)
}
\arguments{
\item{path}{the path you want to open}
}
\description{
open assign path in windows explorer, and default path is
current directory. This function is only designed for windows system.
}
