% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extractId}
\alias{extractId}
\title{extractId}
\usage{
extractId(station, shpfile, formula = ~lon + lat)
}
\arguments{
\item{station}{A data.frame with the station coordinates information}

\item{shpfile}{A character, shape file path.}

\item{formula}{~longitude+latitude}
}
\description{
extractId
}
