% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_sbatch.R
\name{get_sbatch}
\alias{get_sbatch}
\title{get_sbatch}
\usage{
get_sbatch(indir = ".", pattern = "result.*.RDS", outfile = NULL,
  Save = !is.null(outfile))
}
\arguments{
\item{indir}{directory to search rda files}

\item{pattern}{an optional regular expression. Only file names which match 
the regular expression will be returned.}

\item{outfile}{If Save = T, output rda file name.}

\item{Save}{save result in rda file or not?}
}
\description{
Merge the slurm result.
}
