% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwd_clip.R
\name{getwd_clip}
\alias{getwd_clip}
\alias{setwd_clip}
\title{GET or SET working directory}
\usage{
getwd_clip()

setwd_clip()
}
\description{
\itemize{
\item \code{getwd_clip}: get directory path in clipboard, same as getwd function
\item \code{setwd_clip}: set directory path in clipboard, same as setwd function
}
}
\note{
Only works in windows
}
\examples{
\dontrun{
getwd_clip()
setwd_clip()
dir.show()
}
}
\references{
\enumerate{
\item https://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
}
