% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_apply.R
\name{group_apply}
\alias{group_apply}
\alias{group_upperEnvelope}
\title{group_apply}
\usage{
group_apply(x, y, step = 0.2, chunk = NULL, FUN = "mean", ...)

group_upperEnvelope(x, y, step = 0.2, alpha = 0.95)
}
\arguments{
\item{chunk}{split data into nchunks, and apply \code{FUN} in every group}

\item{FUN}{function of mean or median}

\item{...}{others to \code{FUN}}
}
\description{
group_apply

function to separate data to steps of x, obtain 95 quantile value for smooth
}
