% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio.R
\name{import}
\alias{import}
\title{import data to R}
\usage{
import(path, ..., nthreads = 6)
}
\arguments{
\item{nthreads}{Number of threads to use. Default \code{1}.}
}
\description{
import data to R
}
\details{
Support rda, rds, fst, csv, qs
}
