% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{install_gitee}
\alias{install_gitee}
\alias{install_github}
\title{Attempts to install a package directly from gitee.}
\usage{
install_gitee(repo)

install_github(
  repo,
  ref = "master",
  subdir = NULL,
  auth_token = github_pat(quiet),
  host = "api.github.com",
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  repos = getOption("repos"),
  type = getOption("pkgType"),
  ...
)
}
\arguments{
\item{Repository}{address in the format \verb{username/repo[/subdir][@ref|#pull]}.
Alternatively, you can specify subdir and/or ref using the respective parameters (see below);
if both is specified, the values in repo take precedence.}
}
\description{
Attempts to install a package directly from gitee.
}
\examples{
\dontrun{
install_gitee("adv-r/Ipaper")
install_gitee("adv-r/plyr")
}
}
