% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ipaper-package.R
\docType{package}
\name{Ipaper}
\alias{Ipaper}
\alias{_PACKAGE}
\alias{Ipaper-package}
\title{Ipaper}
\description{
Awesome functions in R.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
\keyword{paper}
