% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot2.R
\name{levelplot2}
\alias{levelplot2}
\title{Plot methods for spatial data with attributes}
\usage{
levelplot2(
  formula,
  df,
  SpatialPixel,
  df.mask = NULL,
  brks,
  colors,
  col.rev = FALSE,
  toFactor = FALSE,
  grob = NULL,
  bbox = c(0, 0.5, 0.5, 1),
  xlim = NULL,
  ylim = NULL,
  panel.title = NULL,
  unit = "",
  unit.adj = 0.3,
  pars = NULL,
  stat = list(show = FALSE, name = "RC", loc = c(81.5, 26.5), digit = 1, include.sd =
    FALSE, FUN = weightedMedian),
  sub.hist = TRUE,
  area.weighted = FALSE,
  legend.space = "right",
  layout = NULL,
  colorkey = TRUE,
  legend.num2factor = FALSE,
  interpolate = FALSE,
  lgd.title = NULL,
  sp.layout = NULL,
  NO_begin = 1,
  cex.lgd = 1.3,
  par.settings = opt_trellis_default,
  par.settings2 = list(axis.line = list(col = "white")),
  ...
)
}
\arguments{
\item{formula}{a formula of the form z ~ s1 + s2 | g1 * g2 * ..., where z is a
numeric response, and x, y are numeric values evaluated on a rectangular grid.
g1, g2, ... are optional conditional variables, and must be either factors or
shingles if present.}

\item{df}{data.table object, with columns e.g. lon, lat, and others}

\item{df.mask}{NULL or same length data.table as df, with the columns of \code{mask}
and same group variabes as \code{df}. Mask is used to distinguish significant pixels.
Note that factor levels should be same for grouped variables in \code{df} and \code{df.mask}.}

\item{xlim, ylim}{The limits of x and y}

\item{pars}{parameters controlling hist, e.g. \code{list(title = list(x=77, y=39, cex=1.5),  hist = list(origin.x=77, origin.y=28, A=15, by = 0.4))}.}

\item{sub.hist}{Boolean. If false, hist will be eliminated.}

\item{colorkey}{Boolean or list returned by \code{\link[=get_colorkey]{get_colorkey()}}.}

\item{sp.layout}{NULL or list; see notes below}

\item{NO_begin}{beginning NO of the first panel}

\item{...}{
    Further arguments may be supplied.  Some are processed by
    \code{levelplot} or \code{contourplot}, and those that are
    unrecognized are passed on to the panel function.
  }
}
\description{
Lattice (trellis) plot methods for spatial data with attributes
}
\examples{
data("grid_avhrr")

{
    pars = list(title = list(x=77, y=39, cex=1.5), 
                hist = list(origin.x=77, origin.y=28, A=15, by = 0.4))
    stat = list(show = TRUE, name="mean", loc = c(82.5, 26.5), digit = 1, include.sd = TRUE)
    
    p <- spplot_grid(grid_avhrr, 
                     stat = stat, 
                     pars = NULL, 
                     sub.hist = FALSE,
                     interpolate = FALSE)
    p2 <- levelplot2(X1982 ~ s1+s2, grid_avhrr@data, grid_avhrr)
    write_fig(p , "ex-spplot_grid.png", 9.8, 5)
    write_fig(p2, "ex-spplot_grid2.png", 9.8, 5)
}
}
\seealso{
\code{\link[=spplot_grid]{spplot_grid()}}, \code{\link[sp:spplot]{sp::spplot()}}, \code{\link[lattice:levelplot]{lattice::levelplot()}}
}
