% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listk.R
\name{listk}
\alias{listk}
\title{listk}
\usage{
listk(...)
}
\arguments{
\item{...}{objects, possibly named.}
}
\description{
listk
}
\examples{
a <- 1
b <- 1:2
c <- 1:3
l1 <- listk(a, b, c)
l2 <- listk(a, b, c = 1:3)
l3 <- listk(a = 1, b = 1:2, c = 1:3)
l4 <- listk(1, 1:2, c)
}
