% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{melt_list}
\alias{melt_list}
\title{melt_list}
\usage{
melt_list(list, var.name, na.rm = TRUE, ...)
}
\arguments{
\item{list}{A list object, with the same colnames data.frame in every element.}

\item{var.name}{vector of id variables. Can be integer (variable position) 
or string (variable name). If blank, will use all non-measured variables.}

\item{na.rm}{Boolean}

\item{...}{other parameters to melt}
}
\description{
melt_list
}
