% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_plyr.R
\name{llply}
\alias{llply}
\alias{ldply}
\alias{laply}
\alias{map_simplify}
\title{plyr function in purrr style}
\usage{
llply(.data, .f = NULL, .progress = "none", ...)

ldply(.data, .f = NULL, ...)

laply(.data, .f = NULL, ...)

map_simplify(.data, .f = NULL, ...)
}
\description{
plyr function in purrr style
}
\examples{
x <- list(a = 1:10, beta = exp(-3:3), logic = c(TRUE,FALSE,FALSE,TRUE))
llply(x, mean, .progress = "text")
llply(x, ~mean(.x), .progress = TRUE)
llply(x, quantile, probs = 1:3 / 4)
}
\references{
\enumerate{
\item \url{https://github.com/TylerGrantSmith/purrrgress}
}
}
