% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_list.R
\name{melt_list}
\alias{melt_list}
\alias{melt_tree}
\title{melt_list}
\usage{
melt_list(list, ..., na.rm = TRUE)

melt_tree(x, names, ...)
}
\arguments{
\item{list}{A list object, with the same colnames data.frame in every element.}

\item{...}{other parameters to melt}

\item{na.rm}{Boolean}

\item{var.name}{vector of id variables. Can be integer (variable position)
or string (variable name). If blank, will use all non-measured variables.}
}
\description{
melt_list
}
\examples{
# data.frame
df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
l <- list(a = df, b = df)

melt_list(l, 'id') \%>\% as_tibble()
melt_list(l, 'id' = 1) \%>\% as_tibble()
melt_list(set_names(l, NULL), 'id') \%>\% as_tibble()

# data.table
df <- data.table::data.table(year = 2010, day = 1:3, month = 1, site = "A")
l <- list(a = df, b = df)
melt_list(l, "id")
melt_list(l, id = c("a", "b"))
# int `id` is much more efficient
melt_list(l, id = c(1, 2))
}
\references{
\enumerate{
\item \url{https://stackoverflow.com/questions/15673550/why-is-rbindlist-better-than-rbind}
}
}
