% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart2.R
\name{panel.barchart2}
\alias{panel.barchart2}
\title{panel.barchart2}
\usage{
panel.barchart2(z, subscripts, origin.x = 76, origin.y = 26.5,
  tck = 0.2, A = 15, by = 0.6, box.width = by * 0.85,
  axis.x.text = TRUE, axis.x.text.angle = 90, col,
  fontfamily = "Times", border = "transparent", ntick = 2,
  ylab.offset = 2.5, tick = NULL, w = NULL, text.cex = 1, ...)
}
\arguments{
\item{{[type]}}{perc      \link{description}}

\item{{Number}}{origin.x  \link{description}}

\item{{Number}}{origin.y  \link{description}}

\item{{Number}}{tck       \link{description}}

\item{{Number}}{dx        \link{description}}

\item{{Number}}{A         \link{description}}

\item{{Number}}{by        \link{description}}

\item{{[type]}}{col       Colors}

\item{{[type]}}{box.width \link{description}}

\item{{Number}}{ntick     \link{description}}

\item{{Number}}{text.cex  \link{description}}

\item{{...[type]}}{\link{description}}
}
\value{
{\link{type}}              \link{description}
}
\description{
new panel function for lattice, plot barchart in the bottomleft corner of subplot
}
