% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.spatial.R
\name{panel.spatial}
\alias{panel.spatial}
\title{panel.spatial}
\usage{
panel.spatial(
  x,
  y,
  z,
  subscripts,
  ...,
  contour = FALSE,
  grob = NULL,
  bbox,
  sub.hist = TRUE,
  sp.layout,
  pars,
  class = NULL,
  interpolate = TRUE,
  list.mask = NULL,
  SpatialPixel = NULL,
  density = 1,
  angle = 45,
  par.shade = NULL,
  data.stat = NULL
)
}
\arguments{
\item{par}{position of title, \code{list(title = list(x, y))}}
}
\description{
panel.spatial
}
