% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_lst.R
\name{readxlsx_ToList}
\alias{readxlsx_ToList}
\title{readxlsx_ToList}
\usage{
readxlsx_ToList(fname, show = F, ...)
}
\arguments{
\item{fname}{file name included path going to be read.}
}
\description{
base function, if one excel file hava many sheets, this function will be work
}
\details{
should be cautious that x attribute name can't be NULL should be
}
