% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{first}
\alias{last}
\alias{mutate}
\alias{\%>\%}
\alias{\%<>\%}
\alias{map}
\alias{data.table}
\alias{is.data.table}
\alias{as.data.table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table:as.data.table]{is.data.table}}}

  \item{dplyr}{\code{\link[dplyr:nth]{first}}, \code{\link[dplyr:nth]{last}}, \code{\link[dplyr]{mutate}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{purrr}{\code{\link[purrr]{map}}}
}}

