% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{install_github}
\alias{install_git}
\alias{first}
\alias{last}
\alias{mutate}
\alias{\%>\%}
\alias{\%<>\%}
\alias{set_names}
\alias{map}
\alias{data.table}
\alias{is.data.table}
\alias{as.data.table}
\alias{make_date}
\alias{days_in_month}
\alias{\%dopar\%}
\alias{\%do\%}
\alias{iter}
\alias{rowSums2}
\alias{rowMeans2}
\alias{rowQuantiles}
\alias{icount}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table:as.data.table]{is.data.table}}}

  \item{dplyr}{\code{\link[dplyr:nth]{first}}, \code{\link[dplyr:nth]{last}}, \code{\link[dplyr]{mutate}}}

  \item{foreach}{\code{\link[foreach:foreach]{\%do\%}}, \code{\link[foreach:foreach]{\%dopar\%}}}

  \item{iterators}{\code{\link[iterators]{icount}}, \code{\link[iterators]{iter}}}

  \item{lubridate}{\code{\link[lubridate]{days_in_month}}, \code{\link[lubridate:make_datetime]{make_date}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:aliases]{set_names}}}

  \item{matrixStats}{\code{\link[matrixStats]{rowMeans2}}, \code{\link[matrixStats]{rowQuantiles}}, \code{\link[matrixStats]{rowSums2}}}

  \item{purrr}{\code{\link[purrr]{map}}}

  \item{remotes}{\code{\link[remotes]{install_git}}, \code{\link[remotes]{install_github}}}
}}

