% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_terminal.R, R/main_dt_ddply.R,
%   R/reexports.R, R/set_font.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bold}
\alias{red}
\alias{green}
\alias{underline}
\alias{.}
\alias{\%||\%}
\alias{map}
\alias{mapvalues}
\alias{revalue}
\alias{qsavem}
\alias{qload}
\alias{qsave}
\alias{select}
\alias{rename}
\alias{arrange}
\alias{mutate}
\alias{top_n}
\alias{top_frac}
\alias{as_tibble}
\alias{tribble}
\alias{\%>\%}
\alias{\%<>\%}
\alias{set_names}
\alias{data.table}
\alias{is.data.table}
\alias{fread}
\alias{fwrite}
\alias{as.data.table}
\alias{make_date}
\alias{make_datetime}
\alias{glue}
\alias{days_in_month}
\alias{\%dopar\%}
\alias{\%do\%}
\alias{iter}
\alias{rowSums2}
\alias{rowMeans2}
\alias{rowQuantiles}
\alias{icount}
\alias{install_github}
\alias{install_gitlab}
\alias{install_git}
\alias{edit_r_environ}
\alias{edit_r_profile}
\alias{use_git_ignore}
\alias{use_build_ignore}
\alias{\%->\%}
\alias{\%<-\%}
\alias{showtext_auto}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon:crayon]{bold}}, \code{\link[crayon:crayon]{green}}, \code{\link[crayon:crayon]{red}}, \code{\link[crayon:crayon]{underline}}}

  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table:as.data.table]{is.data.table}}}

  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{as_tibble}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:top_n]{top_frac}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr:reexports]{tribble}}}

  \item{foreach}{\code{\link[foreach:foreach]{\%do\%}}, \code{\link[foreach:foreach]{\%dopar\%}}}

  \item{glue}{\code{\link[glue]{glue}}}

  \item{iterators}{\code{\link[iterators]{icount}}, \code{\link[iterators]{iter}}}

  \item{lubridate}{\code{\link[lubridate]{days_in_month}}, \code{\link[lubridate:make_datetime]{make_date}}, \code{\link[lubridate]{make_datetime}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:aliases]{set_names}}}

  \item{matrixStats}{\code{\link[matrixStats]{rowMeans2}}, \code{\link[matrixStats]{rowQuantiles}}, \code{\link[matrixStats]{rowSums2}}}

  \item{plyr}{\code{\link[plyr:quoted]{.}}, \code{\link[plyr]{mapvalues}}, \code{\link[plyr]{revalue}}}

  \item{purrr}{\code{\link[purrr:reexports]{\%||\%}}, \code{\link[purrr]{map}}}

  \item{qs}{\code{\link[qs:qreadm]{qload}}, \code{\link[qs:qreadm]{qload}}, \code{\link[qs]{qsave}}, \code{\link[qs]{qsavem}}}

  \item{remotes}{\code{\link[remotes]{install_git}}, \code{\link[remotes]{install_github}}, \code{\link[remotes]{install_gitlab}}}

  \item{showtext}{\code{\link[showtext]{showtext_auto}}}

  \item{usethis}{\code{\link[usethis:edit]{edit_r_environ}}, \code{\link[usethis:edit]{edit_r_profile}}, \code{\link[usethis]{use_build_ignore}}, \code{\link[usethis]{use_git_ignore}}}

  \item{zeallot}{\code{\link[zeallot:operator]{\%->\%}}, \code{\link[zeallot:operator]{\%<-\%}}}
}}

