% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ddply.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.}
\alias{mapvalues}
\alias{revalue}
\alias{first}
\alias{last}
\alias{mutate}
\alias{top_n}
\alias{top_frac}
\alias{\%>\%}
\alias{\%<>\%}
\alias{set_names}
\alias{map}
\alias{data.table}
\alias{is.data.table}
\alias{as.data.table}
\alias{make_date}
\alias{make_datetime}
\alias{glue}
\alias{days_in_month}
\alias{\%dopar\%}
\alias{\%do\%}
\alias{iter}
\alias{rowSums2}
\alias{rowMeans2}
\alias{rowQuantiles}
\alias{icount}
\alias{install_github}
\alias{install_gitlab}
\alias{install_git}
\alias{edit_r_environ}
\alias{edit_r_profile}
\alias{use_git_ignore}
\alias{use_build_ignore}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table:as.data.table]{is.data.table}}}

  \item{dplyr}{\code{\link[dplyr:nth]{first}}, \code{\link[dplyr:nth]{last}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:top_n]{top_frac}}, \code{\link[dplyr]{top_n}}}

  \item{foreach}{\code{\link[foreach:foreach]{\%do\%}}, \code{\link[foreach:foreach]{\%dopar\%}}}

  \item{glue}{\code{\link[glue]{glue}}}

  \item{iterators}{\code{\link[iterators]{icount}}, \code{\link[iterators]{iter}}}

  \item{lubridate}{\code{\link[lubridate]{days_in_month}}, \code{\link[lubridate:make_datetime]{make_date}}, \code{\link[lubridate]{make_datetime}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:aliases]{set_names}}}

  \item{matrixStats}{\code{\link[matrixStats]{rowMeans2}}, \code{\link[matrixStats]{rowQuantiles}}, \code{\link[matrixStats]{rowSums2}}}

  \item{plyr}{\code{\link[plyr:quoted]{.}}, \code{\link[plyr]{mapvalues}}, \code{\link[plyr]{revalue}}}

  \item{purrr}{\code{\link[purrr]{map}}}

  \item{remotes}{\code{\link[remotes]{install_git}}, \code{\link[remotes]{install_github}}, \code{\link[remotes]{install_gitlab}}}

  \item{usethis}{\code{\link[usethis:edit]{edit_r_environ}}, \code{\link[usethis:edit]{edit_r_profile}}, \code{\link[usethis]{use_build_ignore}}, \code{\link[usethis]{use_git_ignore}}}
}}

