% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fig.R
\name{showfig}
\alias{showfig}
\title{showfig in external app}
\usage{
showfig(file, pdf.viewer = FALSE)
}
\arguments{
\item{pdf.viewer}{boolean, if set true, pdf viewer will be used mandatorily.}
}
\description{
showfig in external app
}
