% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot_grid.R
\name{spplot_grid}
\alias{spplot_grid}
\title{spplot_grid}
\usage{
spplot_grid(grid, zcols, brks, colors, col.rev = FALSE,
  toFactor = FALSE, sub.hist = TRUE, grob = NULL, bbox = c(0, 0.5,
  0.5, 1), xlim = c(73.5049, 104.9725), ylim = c(25.99376, 40.12632),
  pars = list(title = list(x = 77, y = 39, cex = 1.5), hist =
  list(origin.x = 77, origin.y = 28, A = 15, by = 0.4)), unit = "",
  unit.adj = 0.3, stat = list(show = FALSE, name = "RC", loc = c(81.5,
  26.5)), panel.title = NULL, legend.space = "right",
  layout = layout, colorkey = TRUE, interpolate = TRUE,
  lgd.title = NULL, sp.layout = NULL, ...)
}
\description{
spplot_grid
}
\examples{
\dontrun{
spplot_grid(grid, zcols, )
}

}
