% Generated by roxygen2: do not edit by hand
\name{src_wiley_I}
\alias{src_wiley_I}
\title{srcFUN of WILEY library}
\usage{
src_wiley_I(doi, outdir = "./", srcDownload = TRUE, ...)
}
\arguments{
\item{srcDownload}{If true, it will will download pdf directly, and return 
pdf src. If false, only pdf src returned, without downlaoding pdf.}

\item{...}{other parameters pass to httr::GET}
}
\description{
Journal like GRL, JGR, WRR, HP, all in the database. 
Compared with other srcFUNs, this one is quite complicated. It has to base on 
previous web page request of identidy authentication. So it can't download simply
base on pdf urls, likes other database.
This function was need to further test, whether this function can support 
download with urls
}
