% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contain.R
\name{contain}
\alias{contain}
\title{contain
find assigned pattern variable names}
\usage{
contain(d, pattern = "NDVI|EVI")
}
\arguments{
\item{d}{A data.frame vector, or list}

\item{pattern}{string used to match \code{names(d)}}
}
\description{
contain
find assigned pattern variable names
}
\examples{
df <- data.frame(year = 2010, day = 1:3, month = 1, site = "A")
contain(df, "year|month|day")
}
