% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fig.R
\name{write_fig}
\alias{write_fig}
\title{write_fig}
\usage{
write_fig(p, file = "Rplot.pdf", width = 10, height = 5, res = 300,
  show = T)
}
\description{
show figure relies on SumatraPDF.exe, which needs add to path first.
}
