% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_download.R
\name{write_urls}
\alias{write_urls}
\title{write_urls}
\usage{
write_urls(urls, file)
}
\arguments{
\item{urls}{pdf downloading urls.}

\item{file}{urls are written into file.}
}
\description{
write character vectors of urls into text file, in order to the use of 
subsequent aria2 download
}
