% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_download.R
\name{write_webfile}
\alias{write_webfile}
\title{write_webfile}
\usage{
write_webfile(src, outdir = "./", file = NULL, ...)
}
\arguments{
\item{src}{download link}

\item{outdir}{output directory}

\item{file}{file name}

\item{...}{other parameters to \code{\link[httr]{GET}}}
}
\description{
download web file through web link src
}
