% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_lst.R
\name{writelist_ToXlsx}
\alias{writelist_ToXlsx}
\title{writelist.xlsx}
\usage{
writelist_ToXlsx(x, fname, .progress = "text", rowNames = FALSE, ...)
}
\description{
base function, write list x into fileName with each list in workbook,
write shapefile dbf table into excel *.xlsx
}
\details{
should be cautious that x attribute name can't be NULL should be
}
\examples{
data(hydrodata)
writelist_ToXlsx("hydrodata.xlsx", hydrodata, .progress = "text")
}
